requirements.txt
streamlit>=1.32.0
python-binance>=1.0.19
python-dotenv>=1.0.0
yfinance>=0.2.38
pandas>=2.0.0
numpy>=1.26.0
plotly>=5.18.0
websockets>=12.0
requests>=2.31.0
aiohttp>=3.9.0
cryptography>=42.0.0
fastapi>=0.110.0
uvicorn>=0.27.0
